omics/models: Allow big integers for fasta_offset
Regular integers won't do it with +2GB assemblies

+++ mibios/omics/models.py
@@ -971,7 +971,7 @@ class SequenceLike(Model):
     history = None
     sample = models.ForeignKey(settings.OMICS_SAMPLE_MODEL, **fk_req)
 
-    fasta_offset = models.PositiveIntegerField(
+    fasta_offset = models.PositiveBigIntegerField(
         **opt,
         help_text='offset of first byte of fasta (or similar) header, if there'
                   ' is one, otherwise first byte of sequence',
